Seperated helper functions for easy distinction and testing
